+++ src/main/java/net/frozenblock/lib/FrozenLibMain.java
@@ -51,7 +51,7 @@ import net.frozenblock.lib.particle.FrozenLibParticleTypes;
 import net.frozenblock.lib.registry.FrozenLibRegistries;
 import net.frozenblock.lib.screenshake.api.ScreenShakeManager;
 import net.frozenblock.lib.sound.api.predicate.SoundPredicate;
-import net.frozenblock.lib.spottingicon.api.SpottingIconPredicate;
+import net.frozenblock.lib.spottingicon.api.SpottingIcons;
 import net.frozenblock.lib.tag.api.TagKeyArgument;
 import net.frozenblock.lib.wind.api.WindDisturbanceLogic;
 import net.frozenblock.lib.wind.api.WindManager;
@@ -83,7 +83,7 @@ public final class FrozenLibMain extends FrozenModInitializer {
 		FrozenLibStructureProcessorTypes.init();
 		FrozenLibStructurePoolElementTypes.init();
 		SoundPredicate.init();
-		SpottingIconPredicate.init();
+		SpottingIcons.init();
 		WindDisturbanceLogic.init();
 		FrozenLibDataComponents.init();
 		FrozenLibParticleTypes.init();